requirements.txt
streamlit==1.23.1
torch==2.0.1
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==9.5.0
pandas==2.0.3
SQLAlchemy==2.0.20
